freshness vs. replay protection
From a theoretical perspective I think they are completely different. I think the similarities comes from the use of sequence numbers in most (but not all) replay protections mechanisms. Sequence numbers can be used by the server as a relative measure of when a message was sent on a timescale of the sender.

Suggest to add the following text to the introduction:

"Freshness is a measure of when a message was sent on a timescale of the
recipient. A server that receives a request can either verify that the
request is fresh or determine that it cannot be verified that the request
is fresh. What is considered a fresh is application dependent. Freshness
is completely different from replay protection, but most replay protection
mechanism use a sequence number. Assuming the client is well-behaving, such
a sequence numner that can be used by the server as a relative measure of
when a message was sent on a timescale of the sender."

Maybe the body also need more discussion on replay protection, I will have that in mind while I address Carsten's comments.



#8