delete popular media references.
Achim pointed out that these references are not good. As usual in mainstream media half of the information is incorrect. Delete this for now. Will be replaces with a new text references to companies specializing in DDoS and FBI. High level the severity of the attacks and the bandwidth of the attacks seem to be well known. Details of how the attacks was made and why the stopped are missing.